unify methods in chained_checkboxes_helper, fix "all" query, change accordion open/close & closed by default
