+++ decidim-core/app/services/decidim/resource_search.rb
@@ -24,24 +24,24 @@ module Decidim
     end
 
     # Handle the category_id filter
-    def search_category_id
+    def search_category_ids
       query
         .includes(:categorization)
         .where(decidim_categorizations: { decidim_category_id: category_ids })
     end
 
-    # Handles the scope_id filter. When we want to show only those that do not
-    # have a scope_id set, we cannot pass an empty String or nil because Searchlight
+    # Handles the scope_ids filter. When we want to show only those that do not
+    # have a scope_ids set, we cannot pass an empty String or nil because Searchlight
     # will automatically filter out these params, so the method will not be used.
     # Instead, we need to pass a fake ID and then convert it inside. In this case,
-    # in order to select those elements that do not have a scope_id set we use
+    # in order to select those elements that do not have a scope_ids set we use
     # `"global"` as parameter, and in the method we do the needed changes to search
     # properly.
-    def search_scope_id
-      clean_scope_ids = if scope_id.is_a?(Hash)
-                          scope_id.values
+    def search_scope_ids
+      clean_scope_ids = if scope_ids.is_a?(Hash)
+                          scope_ids.values
                         else
-                          [scope_id].flatten
+                          [scope_ids].flatten
                         end
 
       conditions = []
