+++ decidim-core/app/views/decidim/shared/_chained_check_boxes.html.erb
@@ -1,6 +1,7 @@
   <% if collection.leaf %>
     <% data_sub_checkboxes_id = "#{chained_check_boxes_id}-#{collection.leaf.value}" %>
-    <% is_root_check_box = (collection.leaf.value == "all") ? true : false %>
+    <% hide_node = (hide_node.present? and hide_node == "true") ? true : false %>
+    <% is_root_check_box = (collection.leaf.value == "") ? true : false %>
     <% unless is_root_check_box %>
       <div class="filters__has-subfilters">
     <% end %>
@@ -9,16 +10,17 @@
                          collection.leaf.value,
                          collection.leaf.label,
                          {
-                           "data-sub-checkboxes": data_sub_checkboxes_id
+                           "data-sub-checkboxes": data_sub_checkboxes_id,
+                           is_root_check_box: is_root_check_box
                          }),
-                         collection.leaf.value, nil %>
+                         collection.leaf.value.to_s, nil %>
 
     <% unless is_root_check_box %>
       <% if collection.node.present? %>
         <button type="button" aria-label="show/hide" data-toggle="<%= data_sub_checkboxes_id %> <%= data_sub_checkboxes_id %>-trigger">
           <span id="<%= data_sub_checkboxes_id %>-trigger" data-toggler=".filters__is-open">
-            <span class="filters__icon-is-closed" aria-hidden="true"><%= icon "chevron-bottom" %></span>
-            <span class="filters__icon-is-open" aria-hidden="true"><%= icon "chevron-top" %></span>
+            <span class="filters__icon-is-closed" aria-hidden="true"><%= icon "chevron-top" %></span>
+            <span class="filters__icon-is-open" aria-hidden="true"><%= icon "chevron-bottom" %></span>
           </span>
         </button>
       <% end %>
@@ -26,7 +28,7 @@
     <% end %>
 
     <% if collection.node.present? %>
-      <div id="<%= data_sub_checkboxes_id %>" class="filters__subfilters" <%= %{data-toggler=.hide} unless is_root_check_box %>>
+      <div id="<%= data_sub_checkboxes_id %>" class="filters__subfilters <%= "hide" if hide_node %>" <%= %{data-toggler=.hide} unless is_root_check_box %>>
         <% collection.node.each do |item| %>
           <% if item.tree_node? %>
             <%= render "decidim/shared/chained_check_boxes",
@@ -34,6 +36,7 @@
                         attribute: attribute,
                         collection: item,
                         chained_check_boxes_id: "#{chained_check_boxes_id}-#{item.leaf.value}",
+                        hide_node: "true",
                         options: options %>
           <% else %>
             <%= form.check_box attribute,
